+++ internal/registration/registration.go
@@ -30,7 +30,7 @@ var SUPERBLOCKS_PLUGIN_VERSIONS = map[string][]string{
 	"dynamodb":           {"0.0.7"},
 	"email":              {"0.0.7"},
 	"gcs":                {"0.0.1"},
-	"graphql":            {"0.0.9"},
+	"graphql":            {"0.0.10"},
 	"gsheets":            {"0.0.18"},
 	"kafka":              {"0.0.1"},
 	"javascript":         {"0.0.9"},