+++ CHANGELOG.md
@@ -18,9 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix worker/js helm limits.memory
 - Update shutdown ordering so critical processes are not terminated prior to the webserver
 - Change `v1.security.Resource.config.api` type from `google.protobuf.Value` well known type back to `v1.Api`
-- Added support for enabling `restapi`, `restapiintegration`, `openapi` and `graphql` plugins to return more verbose HTTP response data in result (e.g. response status text, status code, etc.)
-- Added support for configuring whether or not a step using a `restapi`, `restapiintegration`, `openapi` or `graphql` plugin fails if the underlying web request fails
 - Update dependencies to remove critical vulnerabilities in wget and libexpat1
+- Added support for enabling `restapi`, `restapiintegration`, `openapi`, `graphql` and `graphqlintegration` plugins to return more verbose HTTP response data in result (e.g. response status text, status code, etc.)
+- Added support for configuring whether or not a step using a `restapi`, `restapiintegration` or `openapi` plugin fails if the underlying web request fails
+- Added support for configuring whether or not a step using a `graphql` or `graphqlintegration` plugin fails if the GraphQL query returns with errors
 
 ## v1.14.1